# Serialization should not duplicate Tokens and Elements