@mrrodriguez 2016-12-12
>The goal is just to eliminate duplicates by reference, not duplicates by value, correct? I can think of various reasonable scenarios where we'd have internal facts that would be equal but not identical, in some cases lots of them. For example, a fact provided to many different AlphaNodes that all created the same bindings would create distinct Element objects for each matching AlphaNode. [1] I don't have any fundamental objections to optimizing a subset of our use-cases at first but we might need to discuss further off-thread to determine if these changes will solve our (Cerner's) immediate problems.

We talked a bit more on this on the side and I think it is a good idea to consider.  It just would be an unrelated change since we'd then have to start considering how efficient an "equality-style" caching could be done both on serialization and deserialization.

The scope of these changes are centered around just preserving the same number of object references both before and after serialization.  It doesn't consider different object references with equal values.