@EthanEChristian 2016-12-12
Per @WilliamParker's request for performance numbers:

``` Clojure
(require '[clara.rules :as r])
(require '[clara.rules.durability.fressian :as df])
(require '[clara.rules.durability :as d])
(require '[clojure.template :as t])

(defmacro build-records 
  [num]
  `(t/do-template [x] 
                  (defrecord x [~'x ~'y]) 
                  ~@(mapv #(symbol (str "SimpleRecord" %)) (range num))))

(defrecord Output [value])

(build-records 100)

(defmacro make-big-rule
  [num]
  `(r/defrule ~'multi-element-rule
      [SimpleRecord0 (= ~'?x ~'x)]
      [:or ~@(for [x (range 1 num)]
               [(symbol (str "SimpleRecord" x)) (list `= 'x '?x)])]
      ~'=>
      (r/insert! (->Output ~'?x))))

(defrecord LocalMemorySerializer [holder]
  d/IWorkingMemorySerializer
  (serialize-facts [_ fact-seq]
    (reset! holder fact-seq))
  (deserialize-facts [_]
    @holder))

(defn serde-session
  [fired]
  (let [session-baos (java.io.ByteArrayOutputStream.)
        session-serializer (df/create-session-serializer session-baos)
        rulebase-baos (java.io.ByteArrayOutputStream.)
        rulebase-serializer (df/create-session-serializer rulebase-baos)

        holder (atom [])
        mem-serializer (->LocalMemorySerializer holder)

        _ (d/serialize-rulebase fired rulebase-serializer)
        
        _ (println "Serialization Time:")

        _ (time (d/serialize-session-state fired
                                     session-serializer
                                     mem-serializer))

        rulebase-data (.toByteArray rulebase-baos)
        session-data (.toByteArray session-baos)

        rulebase-bais (java.io.ByteArrayInputStream. rulebase-data)
        session-bais (java.io.ByteArrayInputStream. session-data)
        rulebase-serializer (df/create-session-serializer rulebase-bais)
        session-serializer (df/create-session-serializer session-bais)

        restored-rulebase (d/deserialize-rulebase rulebase-serializer)
        _ (println "De-Serialization Time:")
        restored (time (d/deserialize-session-state session-serializer
                                                    mem-serializer
                                                    {:base-rulebase restored-rulebase}))]
  restored))


(make-big-rule 100)


(r/defquery get-outcomes
  []
  [Output (= ?val value)])

(def session (r/mk-session [multi-element-rule get-outcomes]))

(defn build-facts-to-insert
  [num-to-build num-rec-classes]
  (let [builder-fns (mapv #(resolve (symbol (str "->SimpleRecord" %))) (range 1 num-rec-classes))]
  (concat 
    (mapv ->SimpleRecord0 (range num-to-build) (range num-to-build))
    (mapv #(%1 %2 %3) builder-fns (range num-to-build) (range num-to-build)))))

(as-> session ses
      (r/insert-all ses (build-facts-to-insert 1000 100))
      (r/fire-rules ses)
      (let [out (r/query ses get-outcomes)] (println out) ses)
      (serde-session ses))
```

Before:
({:?val 0} {:?val 98} {:?val 97} {:?val 96} {:?val 95} {:?val 94} {:?val 93} {:?val 92} {:?val 91} {:?val 90} {:?val 89} {:?val 88} {:?val 87} {:?val 86} {:?val 85} {:?val 84} {:?val 83} {:?val 82} {:?val 81} {:?val 80} {:?val 79} {:?val 78} {:?val 77} {:?val 76} {:?val 75} {:?val 74} {:?val 73} {:?val 72} {:?val 71} {:?val 70} {:?val 69} {:?val 68} {:?val 67} {:?val 66} {:?val 65} {:?val 64} {:?val 63} {:?val 62} {:?val 61} {:?val 60} {:?val 59} {:?val 58} {:?val 57} {:?val 56} {:?val 55} {:?val 54} {:?val 53} {:?val 52} {:?val 51} {:?val 50} {:?val 49} {:?val 48} {:?val 47} {:?val 46} {:?val 45} {:?val 44} {:?val 43} {:?val 42} {:?val 41} {:?val 40} {:?val 39} {:?val 38} {:?val 37} {:?val 36} {:?val 35} {:?val 34} {:?val 33} {:?val 32} {:?val 31} {:?val 30} {:?val 29} {:?val 28} {:?val 27} {:?val 26} {:?val 25} {:?val 24} {:?val 23} {:?val 22} {:?val 21} {:?val 20} {:?val 19} {:?val 18} {:?val 17} {:?val 16} {:?val 15} {:?val 14} {:?val 13} {:?val 12} {:?val 11} {:?val 10} {:?val 9} {:?val 8} {:?val 7} {:?val 6} {:?val 5} {:?val 4} {:?val 3} {:?val 2} {:?val 1})
**Serialization Time:
"Elapsed time: 2626.406224 msecs"
De-Serialization Time:
"Elapsed time: 748.522282 msecs"**

After:
({:?val 0} {:?val 98} {:?val 97} {:?val 96} {:?val 95} {:?val 94} {:?val 93} {:?val 92} {:?val 91} {:?val 90} {:?val 89} {:?val 88} {:?val 87} {:?val 86} {:?val 85} {:?val 84} {:?val 83} {:?val 82} {:?val 81} {:?val 80} {:?val 79} {:?val 78} {:?val 77} {:?val 76} {:?val 75} {:?val 74} {:?val 73} {:?val 72} {:?val 71} {:?val 70} {:?val 69} {:?val 68} {:?val 67} {:?val 66} {:?val 65} {:?val 64} {:?val 63} {:?val 62} {:?val 61} {:?val 60} {:?val 59} {:?val 58} {:?val 57} {:?val 56} {:?val 55} {:?val 54} {:?val 53} {:?val 52} {:?val 51} {:?val 50} {:?val 49} {:?val 48} {:?val 47} {:?val 46} {:?val 45} {:?val 44} {:?val 43} {:?val 42} {:?val 41} {:?val 40} {:?val 39} {:?val 38} {:?val 37} {:?val 36} {:?val 35} {:?val 34} {:?val 33} {:?val 32} {:?val 31} {:?val 30} {:?val 29} {:?val 28} {:?val 27} {:?val 26} {:?val 25} {:?val 24} {:?val 23} {:?val 22} {:?val 21} {:?val 20} {:?val 19} {:?val 18} {:?val 17} {:?val 16} {:?val 15} {:?val 14} {:?val 13} {:?val 12} {:?val 11} {:?val 10} {:?val 9} {:?val 8} {:?val 7} {:?val 6} {:?val 5} {:?val 4} {:?val 3} {:?val 2} {:?val 1})
**Serialization Time:
"Elapsed time: 519.290918 msecs"
De-Serialization Time:
"Elapsed time: 183.547935 msecs"**