@mrrodriguez 2016-12-13
The performance improvement is looking good in the fabricated case.  However, real cases can come up that result in this sort of rete network too.  We've seen that in practice.