@WilliamParker 2016-12-13
Looks good to me.  Is there anything else we want to do before merging?